refactor to only two main settings mobile/fixed
