--- app/src/main/java/hpsaturn/pollutionreporter/view/SettingsDeviceTools.java
@@ -1,105 +1,0 @@
-package hpsaturn.pollutionreporter.view;
-
-import android.content.SharedPreferences;
-import android.os.Bundle;
-import android.os.Handler;
-
-import androidx.annotation.Nullable;
-import androidx.preference.SwitchPreference;
-
-import com.hpsaturn.tools.Logger;
-
-import java.text.DecimalFormat;
-
-import hpsaturn.pollutionreporter.R;
-import hpsaturn.pollutionreporter.models.CommandConfig;
-import hpsaturn.pollutionreporter.models.GeoConfig;
-import hpsaturn.pollutionreporter.models.ResponseConfig;
-import io.nlopez.smartlocation.SmartLocation;
-
-/**
- * Created by Antonio Vanegas @hpsaturn on 3/31/21.
- */
-public class SettingsDeviceTools extends SettingsBaseFragment{
-
-    public static final String TAG = SettingsDeviceTools.class.getSimpleName();
-
-
-    @Override
-    public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key) {
-
-        if (!onSensorReading) {
-
-            if (key.equals(getString(R.string.key_setting_enable_reboot))) {
-                performRebootDevice();
-            } else if (key.equals(getString(R.string.key_setting_enable_clear))) {
-                performClearDevice();
-            }
-        }
-        else
-            Logger.i(TAG,"skyp onSharedPreferenceChanged because is in reading mode!");
-
-    }
-
-    @Override
-    public void onCreatePreferencesFix(@Nullable Bundle savedInstanceState, String rootKey) {
-        setPreferencesFromResource(R.xml.settings_device_tools, rootKey);
-    }
-
-    @Override
-    protected void refreshUI() {
-
-    }
-
-    @Override
-    protected void onConfigRead(ResponseConfig config) {
-
-    }
-
-
-    private void performRebootDevice() {
-        SwitchPreference rebootSwitch = findPreference(getString(R.string.key_setting_enable_reboot));
-        if (!rebootSwitch.isChecked()) {
-            if(snackBar!=null)snackBar.dismiss();
-        } else {
-            snackBar = getMain().getSnackBar(R.string.bt_device_reboot, R.string.bt_device_reboot_action, view -> {
-                rebootSwitch.setChecked(false);
-                sendSensorReboot();
-                getMain().showSnackMessageSlow(R.string.msg_device_reboot);
-                Handler handler = new Handler();
-                handler.postDelayed(() -> getMain().finish(), 3000);
-            });
-            snackBar.show();
-        }
-    }
-
-    private void sendSensorReboot() {
-        CommandConfig config = new CommandConfig();
-        config.cmd = getSharedPreference(getString(R.string.key_setting_wmac));
-        config.act = "rbt";
-        sendSensorConfig(config);
-    }
-
-    private void performClearDevice() {
-        SwitchPreference clearSwitch = findPreference(getString(R.string.key_setting_enable_clear));
-        if (!clearSwitch.isChecked()) {
-            if(snackBar!=null)snackBar.dismiss();
-        } else {
-            snackBar = getMain().getSnackBar(R.string.bt_device_clear, R.string.bt_device_clear_action, view -> {
-                clearSwitch.setChecked(false);
-                CommandConfig config = new CommandConfig();
-                config.cmd = getSharedPreference(getString(R.string.key_setting_wmac));
-                config.act = "cls";
-                getMain().showSnackMessageSlow(R.string.msg_device_clear);
-                sendSensorConfig(config);
-                clearSharedPreferences();
-                Handler handler = new Handler();
-                handler.postDelayed(() -> getMain().finish(), 3000);
-            });
-            snackBar.show();
-        }
-    }
-
-
-
-}
