+++ app/src/main/java/hpsaturn/pollutionreporter/view/SettingsMobileStation.java
@@ -1,0 +1,214 @@
+package hpsaturn.pollutionreporter.view;
+
+import android.content.SharedPreferences;
+import android.os.Bundle;
+import android.os.Handler;
+
+import androidx.annotation.Nullable;
+import androidx.preference.ListPreference;
+import androidx.preference.SwitchPreference;
+
+import com.hpsaturn.tools.Logger;
+
+import hpsaturn.pollutionreporter.R;
+import hpsaturn.pollutionreporter.models.CommandConfig;
+import hpsaturn.pollutionreporter.models.ResponseConfig;
+import hpsaturn.pollutionreporter.models.SampleConfig;
+import hpsaturn.pollutionreporter.models.SensorType;
+
+/**
+ * Created by Antonio Vanegas @hpsaturn on 3/31/21.
+ */
+public class SettingsMobileStation extends SettingsBaseFragment{
+
+    public static final String TAG = SettingsMobileStation.class.getSimpleName();
+
+
+    @Override
+    public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key) {
+
+        if (!onSensorReading) {
+
+            if (key.equals(getString(R.string.key_setting_stime))) {
+                validateSensorSampleTime();
+            }
+            else if (key.equals(getString(R.string.key_setting_dtype))) {
+                sendSensorTypeConfig();
+            }
+            else if (key.equals(getString(R.string.key_setting_enable_reboot))) {
+                performRebootDevice();
+            }
+            else if (key.equals(getString(R.string.key_setting_enable_clear))) {
+                performClearDevice();
+            }
+        }
+        else
+            Logger.i(TAG,"skyp onSharedPreferenceChanged because is in reading mode!");
+
+    }
+
+    @Override
+    public void onCreatePreferencesFix(@Nullable Bundle savedInstanceState, String rootKey) {
+        setPreferencesFromResource(R.xml.settings_mobile_station, rootKey);
+    }
+
+    @Override
+    protected void refreshUI() {
+        updateStimeSummary();
+
+    }
+
+    @Override
+    protected void onConfigRead(ResponseConfig config) {
+        Logger.d(TAG, "[Config] onConfigCallBack");
+
+        boolean notify_sync = false;
+
+        if (!getStimeFormat(config.stime).equals(getStimeFormat(getCurrentStime()))){
+            notify_sync = true;
+        }
+        if (config.stype != getSensorType()) {
+            if (config.stype < 0) updateSensorTypeSummary(0);
+            else updateSensorTypeSummary((config.stype));
+        }
+
+        setStatusSwitch(true);
+
+        if (notify_sync) {
+            saveAllPreferences(config);
+            updateStatusSummary(true);
+            updatePreferencesSummmary(config);
+            Logger.v(TAG, "[Config] notify device sync complete");
+            getMain().showSnackMessage(R.string.msg_sync_complete);
+        }
+
+    }
+
+    /***********************************************************************************************
+     * Sample time handlers
+     **********************************************************************************************/
+
+    private void validateSensorSampleTime() {
+        Logger.v(TAG, "[Config] validating->" + getString(R.string.key_setting_stime));
+        if (getCurrentStime() >= 5) {
+            saveSensorSampleTime(getCurrentStime());
+        }
+    }
+
+    private void saveSensorSampleTime(int time) {
+        Logger.v(TAG, "[Config] sending sensor time: "+time);
+        SampleConfig config = new SampleConfig();
+        config.stime = time;
+        updateSummary(R.string.key_setting_stime,getStimeFormat(config.stime));
+        sendSensorConfig(config);
+    }
+
+    private int getCurrentStime() {
+        try {
+            return Integer.parseInt(getSharedPreference(getString(R.string.key_setting_stime)));
+        } catch (NumberFormatException e) {
+            e.printStackTrace();
+        }
+        return 0;
+    }
+
+    private void resetStimeValue(int stime) {
+        saveSharedPreference(R.string.key_setting_stime, "" + stime);
+        updateStimeSummary();
+    }
+
+    /***********************************************************************************************
+     * Sensor type section
+     **********************************************************************************************/
+
+    private void sendSensorTypeConfig() {
+        Logger.v(TAG, "[Config] validating->" + getString(R.string.key_setting_dtype));
+        getMain().showSnackMessage(R.string.msg_reboot_manually);
+        SensorType config = new SensorType();
+        config.stype = getSensorType();
+        sendSensorConfig(config);
+    }
+
+    private void updateSensorTypeSummary(int value) {
+        ListPreference sizePreference = findPreference(getString(R.string.key_setting_dtype));
+        sizePreference.setValueIndex(value);
+    }
+
+    private int getSensorType() {
+        try {
+            return Integer.parseInt(getSharedPreference(getString(R.string.key_setting_dtype)));
+        } catch (NumberFormatException e) {
+            e.printStackTrace();
+        }
+        return 0;
+    }
+
+    String getStimeFormat(int time){
+        return "" + time + " seconds";
+    }
+
+    private void updateStimeSummary() {
+        updateSummary(R.string.key_setting_stime,getStimeFormat(getCurrentStime()));
+    }
+
+    private void updatePreferencesSummmary(ResponseConfig config) {
+        if(config.stime>0)updateSummary(R.string.key_setting_stime, getStimeFormat(config.stime));
+    }
+
+    /***********************************************************************************************
+     * Sensor tools
+     **********************************************************************************************/
+
+    private void performRebootDevice() {
+        SwitchPreference rebootSwitch = findPreference(getString(R.string.key_setting_enable_reboot));
+        if (!rebootSwitch.isChecked()) {
+            if(snackBar!=null)snackBar.dismiss();
+        } else {
+            snackBar = getMain().getSnackBar(R.string.bt_device_reboot, R.string.bt_device_reboot_action, view -> {
+                rebootSwitch.setChecked(false);
+                sendSensorReboot();
+                getMain().showSnackMessageSlow(R.string.msg_device_reboot);
+                Handler handler = new Handler();
+                handler.postDelayed(() -> getMain().finish(), 3000);
+            });
+            snackBar.show();
+        }
+    }
+
+    private void sendSensorReboot() {
+        CommandConfig config = new CommandConfig();
+        config.cmd = getSharedPreference(getString(R.string.key_setting_wmac));
+        config.act = "rbt";
+        sendSensorConfig(config);
+    }
+
+    private void performClearDevice() {
+        SwitchPreference clearSwitch = findPreference(getString(R.string.key_setting_enable_clear));
+        if (!clearSwitch.isChecked()) {
+            if(snackBar!=null)snackBar.dismiss();
+        } else {
+            snackBar = getMain().getSnackBar(R.string.bt_device_clear, R.string.bt_device_clear_action, view -> {
+                clearSwitch.setChecked(false);
+                CommandConfig config = new CommandConfig();
+                config.cmd = getSharedPreference(getString(R.string.key_setting_wmac));
+                config.act = "cls";
+                getMain().showSnackMessageSlow(R.string.msg_device_clear);
+                sendSensorConfig(config);
+                clearSharedPreferences();
+                Handler handler = new Handler();
+                handler.postDelayed(() -> getMain().finish(), 3000);
+            });
+            snackBar.show();
+        }
+    }
+
+    /***********************************************************************************************
+     * Update Preferences methods
+     **********************************************************************************************/
+
+    private void saveAllPreferences(ResponseConfig config) {
+        resetStimeValue(config.stime);
+    }
+
+
+}
