+++ app/src/main/res/xml/settings_general.xml
@@ -12,18 +12,9 @@
             android:summary="What variable is shown in each chart"
             android:title="Variables filter"
             app:useSimpleSummaryProvider="true" />
-
-        <EditTextPreference
-            android:inputType="textUri"
-            app:defaultValue="/sdcard/canairio"
-            app:enabled="false"
-            app:key="@string/key_setting_sdpath"
-            app:summary="/sdcard/canairio"
-            app:title="Auto Save Directory" />
-
     </PreferenceCategory>
 
-    <PreferenceCategory >
+    <PreferenceCategory>
         <SwitchPreference
             app:defaultValue="false"
             app:icon="@android:drawable/ic_lock_power_off"
@@ -33,41 +24,25 @@
     </PreferenceCategory>
     <PreferenceCategory
         android:dependency="@string/key_device_status"
-        app:title="Device Settings">
+        app:title="CanAirIO Device Settings">
 
-        <EditTextPreference
-            android:inputType="number"
-            app:defaultValue="5"
-            app:icon="@android:drawable/ic_menu_save"
-            app:key="@string/key_setting_stime"
-            app:summary="@string/summary_stime"
-            app:title="Sample Time Interval" />
-
-        <ListPreference
-            android:defaultValue="0"
-            android:entries="@array/pref_dtype_entries"
-            android:entryValues="@array/pref_dtype_values"
-            android:key="@string/key_setting_dtype"
-            android:title="Sensor type"
-            app:useSimpleSummaryProvider="true" />
-
+        <Preference
+            android:dependency="@string/key_device_status"
+            android:fragment="hpsaturn.pollutionreporter.view.SettingsMobileStation"
+            app:summary="Bluetooth mobile station settings"
+            app:title="Mobile Station"></Preference>
 
         <Preference
             android:dependency="@string/key_device_status"
             android:fragment="hpsaturn.pollutionreporter.view.SettingsFixedStation"
             app:summary="WiFi fixed station settings"
-            app:title="Fixed Station">
-        </Preference>
+            app:title="Fixed Station"></Preference>
 
-        <Preference
-            android:dependency="@string/key_device_status"
-            android:fragment="hpsaturn.pollutionreporter.view.SettingsDeviceTools"
-            app:summary="Some device actions and settings"
-            app:title="CanAirIO device tools">
-        </Preference>
     </PreferenceCategory>
 
-    <PreferenceCategory app:title="">
+    <PreferenceCategory
+        android:dependency="@string/key_device_status"
+        app:title="">
         <Preference
             app:icon="@android:drawable/ic_popup_sync"
             app:key="@string/key_device_info"
@@ -75,8 +50,7 @@
             app:title="Device info:" />
     </PreferenceCategory>
 
-    <PreferenceCategory
-        app:title="">
+    <PreferenceCategory app:title="">
         <Preference
             app:icon="@android:drawable/ic_menu_send"
             app:key="@string/key_send_feedback"
