+++ app/src/main/res/xml/settings_mobile_station.xml
@@ -11,6 +11,35 @@
             app:title="Device Status:" />
     </PreferenceCategory>
 
+    <PreferenceCategory
+        android:dependency="@string/key_device_status"
+        app:title="CanAirIO Device Settings">
+
+        <EditTextPreference
+            android:inputType="number"
+            app:defaultValue="5"
+            app:icon="@android:drawable/ic_menu_save"
+            app:key="@string/key_setting_stime"
+            app:summary="@string/summary_stime"
+            app:title="Sample Time Interval" />
+
+        <ListPreference
+            android:defaultValue="0"
+            android:entries="@array/pref_dtype_entries"
+            android:entryValues="@array/pref_dtype_values"
+            android:key="@string/key_setting_dtype"
+            android:title="Sensor type"
+            app:useSimpleSummaryProvider="true" />
+
+        <EditTextPreference
+            android:inputType="textUri"
+            app:defaultValue="/sdcard/canairio"
+            app:enabled="false"
+            app:key="@string/key_setting_sdpath"
+            app:summary="/sdcard/canairio"
+            app:title="Auto Save Directory" />
+
+    </PreferenceCategory>
 
     <PreferenceCategory
         android:dependency="@string/key_device_status"
